Rename README.md to week-05-o-neil.md
According to "Weapons of Math Destruction", by Cathy O'Neil, a WMD in short is a scenario that is predicted based on prior data. It foreshadows how a particular event may go based on mathematical, or statistical data collected from the past of a given scenario. For example, as aforementioned in the article, a baseball game. In the baseball game he mentioned, Ted Williams was notoriously known in a doubleheader competition for his aggressive line drives. In order to avoid continuous defeat, an opposing team player, Lou Boudreau, studied where he Williams was mostly hitting the ball, which happened to be right field. Based off of this knowledge, he, and his teammates were able to catch more of his plays, proving this to me an example WMD. 

A personal example of this occurrence would be how I choose to do my schoolwork. In previous years, I've had to deal with multiple large projects, and assignments. Oftentimes, there were very tight time constraints on some of them. In an effort to better manage handling the more time consuming tasks, I took in the prompt of each assignment, and weighed the necessary time needed to complete them. Based off of data I collected on how rigorous the task was, or detailed,  I was able to make a WMD of the situation as a whole. I deduced that due to the greater amount of requirements needed for the larger projects, I would need to dedicate more time to those, which led me to start with this assignment first. Subsequently, I would then choose to spend more time on these type of assignments, and continue to chip away at the others of lesser requirements/importance slowly. 

As a result of taking this approach, I found that more times than not, this was very  effective in helping me meet the deadline for the larger project at hand successfully. Although it seems like a good strategy off hand, it wasn't always the best one. More times than not, this approach has worked for me, but at times it has failed me. Sometimes, different phases of the project took longer than expected, and I risked nearly missing, or missing the deadline for it. This proved to be very difficult. However, my misunderstanding, or lack of better method for navigating through part of the project, at times was out of my control without experimentation. Also, a lack of experience also proved to be detrimental in this case at times. Just as a WMD such as this can prove, every situation is different from case to case, and is never certain, but probable.